requirements.txt
python-dotenv
requests
pydantic
numpy
sentence-transformers
umap-learn
hdbscan
fastapi
uvicorn
